+++ addons/gut/gut_cmdln.gd
@@ -262,7 +262,7 @@ func _run_gut():
 
 			var runner = GutRunner.instantiate()
 
-			runner.set_cmdln_mode(true)
+			runner.ran_from_editor = false
 			runner.set_gut_config(_gut_config)
 
 			get_root().add_child(runner)