+++ addons/gut/gut_config.gd
@@ -48,16 +48,6 @@ var default_options = {
 	gut_on_top = true,
 }
 
-var default_panel_options = {
-	font_name = 'CourierPrime',
-	font_size = 16,
-	output_font_name = 'CourierPrime',
-	output_font_size = 30,
-	hide_result_tree = false,
-	hide_output_text = false,
-	hide_settings = false,
-	use_colors = true
-}
 
 var options = default_options.duplicate()
 var json = JSON.new()
@@ -165,13 +155,18 @@ func write_options(path):
 	return result
 
 
+# consistent name
+func save_file(path):
+	write_options(path)
+
+
 func load_options(path):
 	return _load_options_from_config_file(path, options)
 
 
-func load_panel_options(path):
-	options['panel_options'] = default_panel_options.duplicate()
-	return _load_options_from_config_file(path, options)
+# consistent name
+func load_file(path):
+	return load_options(path)
 
 
 func load_options_no_defaults(path):