+++ addons/gut/utils.gd
@@ -1,3 +1,4 @@
+@tool
 class_name GutUtils
 # ------------------------------------------------------------------------------
 # Description
@@ -35,6 +36,9 @@ const TEST_STATUSES = {
 	FAILED = 'fail',
 	PASSED = 'pass'
 }
+
+static var avail_fonts = ['AnonymousPro', 'CourierPrime', 'LobsterTwo', 'Default']
+
 
 # This is a holdover from when GUT was making a psuedo autoload.  It would add
 # an instance of this class to the tree with a name and retrieve it when
@@ -190,9 +194,9 @@ var CollectedScript = load('res://addons/gut/collected_test.gd')
 var GutScene = load('res://addons/gut/GutScene.tscn')
 
 # Source of truth for the GUT version
-var version = '9.1.2'
+var version = '9.2.0'
 # The required Godot version as an array.
-var req_godot = [4, 1, 0]
+var req_godot = [4, 2, 0]
 
 
 # ------------------------------------------------------------------------------