+++ scenes/TestShortcuts.gd
@@ -1,20 +1,6 @@
 extends Node2D
 
-
-# Declare member variables here. Examples:
-# var a = 2
-# var b = "text"
-
 var count = 1
-
-# Called when the node enters the scene tree for the first time.
-func _ready():
-	pass # Replace with function body.
-
-
-# Called every frame. 'delta' is the elapsed time since the previous frame.
-#func _process(delta):
-#	pass
 
 
 func _on_ShortcutButton_changed():
@@ -25,3 +11,7 @@ func _on_ShortcutButton_changed():
 func _on_Button_pressed():
 	print('hello world ', count)
 	count += 1
+
+
+func _on_show_dialog_pressed():
+	$BottomPanelShortcuts.popup_centered()